requirements.txt
Flask==3.0.0
opencv-python==4.8.1.78
mediapipe==0.10.8
Pillow==9.5.0
